refactor: garbled_groth16 input
